Cria módulo normalizer
